pyproject.toml
[project]
name = "EasyKit"
version = "3.2.1"
description = "Automate tasks, simplify Windows maintenance, and eliminate repetitive processes with this Python-powered toolkit. Contributions are welcome to enhance its functionality!"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "build>=1.2.2.post1",
    "click>=8.2.0",
    "colorama>=0.4.6",
    "packaging>=25.0",
    "platformdirs>=4.3.8",
    "python-dotenv>=1.1.0",
    "pywin32>=310",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "setuptools>=80.4.0",
    "twine>=6.1.0",
    "wheel>=0.45.1",
]

[tool.setuptools.packages.find]
include = ["easykit*", "images*", "docs*"]
